requirements.txt
torch>=2.6.0
torchvision>=0.21.0
# NumPy 2.0 removed np.complex_ etc.; tensorflow_graphics / parts of the stack still expect 1.x
numpy>=1.26.4,<2.0
scipy>=1.16.0

accelerate>=1.10.0
peft>=0.11.0
safetensors>=0.6.0
timm>=0.9.10
einops>=0.8.0
diffusers>=0.30.0

gymnasium>=0.29.0
robosuite==1.4.1
mujoco>=3.3.0
sapien>=3.0.0

hydra-core>=1.3.0
omegaconf>=2.3.0
draccus>=0.8.0
tyro>=0.9.0
pydantic>=2.0

h5py>=3.14.0
datasets>=4.0.0
pyarrow>=21.0.0

wandb>=0.21.0
tensorboard>=2.15.0
rich>=13.0.0

imageio>=2.37.0
imageio-ffmpeg>=0.6.0
opencv-python>=4.11.0
matplotlib>=3.10.0
tqdm>=4.67.0
ray==2.49.0
PyYAML>=6.0.0

easydict
bddl
tensorflow
tensorflow-graphics
